fix: resolve 32-bit uint overflow on arm and add multi-arch CI check
    Prevent constant overflow compile errors on 32-bit architectures (such as
    linux/arm) and automate multi-architecture cross-validation across Makefile,
    CI, and release workflows.

    Key changes:
    - Type conversion: In toInt64 (resource_manifest.go), promote uint to
      uint64 before comparing with math.MaxInt64. On 32-bit architectures,
      uint is 32 bits, which previously caused the untyped constant
      math.MaxInt64 (2^63 - 1) to overflow uint at compile time.
    - Tests: Add 32-bit uint boundary checks in resource_manifest_test.go.
    - Makefile: Add cross-check target validating go vet and cross-compilation
      across all release targets (linux/amd64, linux/arm, linux/arm64,
      windows/amd64, windows/arm64, darwin/amd64, darwin/arm64); include in
      make all.
    - CI/Release: Add cross-check pre-flight verification to the release
      workflow (.github/workflows/release.yml) and introduce a dedicated CI
      workflow (.github/workflows/ci.yml) for PRs and pushes to main.

+++ kubefu/resource_manifest.go
@@ -1002,7 +1002,7 @@ func toInt64(v any) (int64, bool) {
 	case int64:
 		return n, true
 	case uint:
-		if n <= math.MaxInt64 {
+		if uint64(n) <= math.MaxInt64 {
 			return int64(n), true
 		}
 	case uint8:
